Moved main view into it's own file.
