+++ ViewModels/MainWindowViewModel.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using Avalonia.Media;
+using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
 using Paints.Models;
 
@@ -9,28 +10,8 @@ namespace Paints.ViewModels;
 
 public partial class MainWindowViewModel : ViewModelBase
 {
-    private List<Paint> _paints = [
-        new Paint("Whimsical White", "Eris", Colors.White),
-        new Paint("Bulbous Blue", "Eris", Colors.Blue),
-        new Paint("Silly Salamander", "Eris", Colors.LightPink),
-        new Paint("Goose Gray", "Eris", Colors.Gray)
-    ];
+    private ListPageViewModel _listPageViewModel = new ListPageViewModel();
 
-    public IEnumerable<PaintViewModel> Paints => _paints.Select(PaintViewModelFactory);
     
-    private void SelectPaint(PaintViewModel paint)
-    {
-        Console.WriteLine(paint.Name + " selected!");
-    }
-
-    private PaintViewModel PaintViewModelFactory(Paint paint)
-    {
-        var viewModel = new PaintViewModel(paint);
-        viewModel.SelectCommand = new RelayCommand(() =>
-        {
-             SelectPaint(viewModel);
-        });
-
-        return viewModel;
-    }
+    public ViewModelBase CurrentPage => _listPageViewModel;
 }
