+++ Views/ListPageView.axaml.cs
@@ -1,0 +1,13 @@
+using Avalonia;
+using Avalonia.Controls;
+using Avalonia.Markup.Xaml;
+
+namespace Paints.Views;
+
+public partial class ListPageView : UserControl
+{
+    public ListPageView()
+    {
+        InitializeComponent();
+    }
+}
